toolsdb: Make socket a parameter so new servers might work right
The existing servers probably need gentility, but as I'm building a new
one, it seems that we should not be using a "socket" setting that always
fails after a reboot. For now, the default is the broken one so things
remain the same on existing servers. Fix that as we shut them down.

Also removing the unused (and goofy as hell, in cloud) ::passwords
mention.

Bug: T266587
Change-Id: I08dc726693349895fd84238449433660e84a8bed