site: introduce mwdebug1003 as debug server on buster
This will introduce a new ganeti VM on Debian buster
using the MediaWiki appserver role, as previously
applied on testvm1001.

Based on that experience I highly recommend in this case to
not immediately apply the prod puppet role but to first add it with
the "insetup" role, create it, run puppet and then switch
the role.

And before that can happen there also need to be new mcrouter
certs for it.

https://wikitech.wikimedia.org/wiki/Memcached_for_MediaWiki/mcrouter#Generate_certs_for_a_new_host

But as it is this can be merged before or after creating the VM
using the cookbook.

Bug: T245757
Change-Id: I9653c9e99fdedecbaf1c39d279234ca23b9cdd44