kartotherian: Don't page SREs on failure
Why:
SREs are being paged about a service whose ownership and level of
support is currently unclear. Pages triggered for it always end up in
the same argument about level of support and the infrastructure
requiring work to happen to become sustainable. Those discussions
actively lead to lower morale and as such are counter productive as well
as interrupting SREs from other issues they are working on.

Switching to an alert but not page model while the ownership of maps is
being sorted out will hopefully address the above.

What:
Switch from critical true to false. This will stop pages to SREs but
retain IRC alerts

Bug: T267339
Change-Id: I7f77fd9a1c8d49f3e23dc4068f8a63f10f1f6ad9