cloud-vps: Change NFS mounts to default to false
An error in the puppet enc can cause NFS to be mounted across an entire
project. However, errors do not cause puppet to unmount NFS shares. The
reason is that puppet does not yet have a way to remove shares unless
changes are made to share definitions, not based on the mount_nfs hiera
key. While users can easily declare "mount_nfs: true" across their whole
project, you have to rebuild or go through a lot of trouble to undo what
"mount_nfs: true" does on hosts you do not want NFS on.

This will fix the problem, but it will require a bit of user education
regarding how to get NFS to mount on a new VM.

Bug: T262350
Bug: T267082
Change-Id: Ie66f88af078beab8cee532db42131d35b84c9ad0