postgres: change max replication connections to nodes + 6
We can safely add 3 new replicas without majorly impacting performance
on the master based on the experiments in eqiad. Currently our setup
allows for the pooling of a single node, as it will add 1 to the
len(nodes) connections we currently allow. This is a bit restrictive,
so open up some room to allow for a few more concurrent replicas to be
created.

Bug: T266820
Change-Id: Ie3ef9d1e6190fd86b924c617e116da787219bb2f