P:tendril::webserver: migrate to ensure_packages
Ubdate the code to use ensure_packages instead of require_packages,
please check the bug refrenrence for more information on the finer
details.  The main point is that we now have to explicitly define an
resources which depend on ensured_packages.

I have also performed a minor refactor.

Bug: T266479
Hosts: P:tendril::webserver
Change-Id: If3826ed1162b3d354a8f2da892f2a89a4e8822df