kerberos: add dns_canonicalize_hostname = false to clients
The dns_canonicalize_hostname option defaults to true on all krb
versions prior 1.19 (we run on 1.17 on Buster), but it is set to
false after it. Usually this is not a problem, but for corner cases
like DNS CNAMEs it could lead to some clients failing to authenticate
to some services. A recent example for Analytics is the following:
we added a CNAME for an-test-coord1001 called 'analytics-test-hive',
in order to avoid hardcoding the former in every client config.
We also changed the service principal for the hive daemon from
'hive/an-test-coord1001.eqiad.wmnet@WIKIMEDIA' to
'hive/analytics-test-hive.eqiad.wmnet@WIKIMEDIA', and tested various
clients. The result was interesting: java-based clients don't pay
attention to defaults in /etc/krb.conf, so the new CNAME worked fine,
but clients using c-based libraries (SASL mostly) do care about
the config files, and don't work if dns_canonicalize_hostname=true since
a forward DNS resolution of the CNAME causes the client to get a token
for the wrong service principal.
We have two options:
- we set dns_canonicalize_hostname=true in the krb.conf, so future
  versions of krb do not change it, and we test java-based clients
  with the new settings.
- we assume that the dns_canonicalize_hostname=false setting in 1.19
  was done for a good reason and we keep the current auth scheme,
  hence not asking clients to resolve DNS CNAMEs.

More info in https://web.mit.edu/kerberos/krb5-devel/doc/admin/princ_dns.html

Bug: T257412
Change-Id: Id07f56d876fbe42db5626259ef04205c15c5a2e6