prometheus::mysqld_exporter::instance: drop arguments paramteter
The arguments paramter has not been used in production for some time
and its inclusion as an option complicates the argument handling
logic. Suggest we drop this p[aramter and ad back late if its needed in
the future

This CR drops the arguments parameter and refactors the argument
handling.

Hosts: C:prometheus::mysqld_exporter::common
Bug: T256972
Change-Id: Iec2815a4a05a423d5e1bd6b0d8ee31f4bc451a16