elastic: Turn on adaptive replica selection in elastic 6
This is the default in elastic 7, in elastic 6 it was still new and
not yet the default. In review there were no changes made to this
functionality between our release and when it became the default, and
has been enabled in the transient settings of the search platform
eqiad clusters for the past few months.

Bug: T259539
Change-Id: I1b35867c7a57fe880d164e7ec662336fcb311cdf