Update network topology for Hadoop worker nodes
The SRE dc-ops team is going to move some servers from one rack to the
other one to free space for 10g ports. Some of our 1g-based workers
are in the list, and this change updates the network topology to reflect that.
I think it is fine to merge and update the namenode beforehand (so blocks will be
replicated according to the new topology).

Bug: T267065
Change-Id: Ice5550b1f3e7b07c8984eb78e65a54899fbbae91