releases::mediawiki: do not attempt to restart masked service
The wmf_auto_restart_* is nowadays a systemd timer and we
end up with Icinga alerts about a broken systemd state on
some hosts.

This happens where a service is configured as absent or masked
on a standby/passive host while it's running/enabled on an
active host.

Then the systemd timer fails because it tries to restart a service
that is simply "not present" from systemd point of view.

So since there can't be restarting of service that does not
exist and it creates monitoring noise make sure we only
apply the restart if the service is not masked.

The fact this defaults to masked was once a special request
for jenkins.

Change-Id: Ie7dd67b05f03be98f8f0e2a20e621f9f4e42cc2d