admin: create peek-admins group and apply on peek role
A follow-up to T265922 where we handed over access to the
VM using the "peek" role (a tool used by the secteam) which
was previously maintained by Chase.

Chase had global root but secteam members have not. So we did
give them SSH access but they can't run anything as a privileged user
so far.

Currently the need is to make it possible for Reedy and Scott to
take over maintenance of the peek tool.

For that I originally suggested a temp. root access (this is a dedicated
machine for this tool only anyways) but as Keith rightfully
points out we can just go with "any command as peek user" probably.

So let's see if that is sufficient.

Bug: T265922
Change-Id: Ib0bf974e6302ec94af6d6fdaa6d0f2baa869c657