bacula::client: notify service after installing certs
Puppet currently tries to start the bacula-fd servie before the puppet
certificates have been exposed.  Fruther the service is not notified if
theses certificates change e.g. if they are renewed

This CR updates the code so that the expose_puppet_crts type notifies
the bacula-fd service to ensure it is correctly refreshed when certs are
updated.  Further notify implies before so the certs should be installed
before puppet attempts to start the service on the first run

Hosts: cumin2001.codfw.wmnet
Bug: T256454
Change-Id: Id1269d3db079637216c6f6ff2b11af9e56f6e51f