authdns: copy geoip data rather than symlink
Symlinks defeat ev_stat monitoring of the configured path, but
puppet's file resources can be used to instead just copy the data
into place.  This should cause gdnsd to actually notice GeoIP
database changes in a timely fashion!

In theory it would suffice for production's purposes to
reconfigure gdnsd to look directly at the absolute path in
/usr/share, but we need the currently-configured relative path in
order for CI checks to work in non-production.

Bug: T252577
Change-Id: I82da5d46cd3525f8c0a055c98aabe20a3c09ee53