hieradata: enable onhost memcached on api and app canaries
Moving forward with enabling onhost memcached
on api and app canaries. All keys prefixed 
with /eqiad/mw-with-onhost-tier will be fetched from the 
main memcached cluster but will be cached (max TTL 10s) 
and read from a memcached instance running on the 
host itself.

mwdebug1003 is excluded since it will be used for the
mediawiki to buster upgrade.

Bug: T244340
Change-Id: I3c57c57dec6e143f5e29b2bfbb1a77ee0e5920d8