idp: update idp_primary/idp_failover
As per dns the primary idp server is currently idp2001.  This got missed
last time we failed over and is not the first time it has been missed.
I think we shoul at the very least have some icinga check for this but
perhaps we could base the idp_primary/idp_failover configueration based
on the result of

 dig CNAME idp.wikimedia.org [@ns0.wikimedia.org]?

Hosts: O:idp
Change-Id: Iaeb568a4b34db1d0cbc00925d88e98712dc5cdc2