kerberos::exec: enable kerberos by default
This patch removes redundant "enable kerberos" configs that don't
make sense anymore in production. It also adds a flag to avoid
using kerberos in labs/cloud, to ease testing if needed (the chance
that we'll bootstrap a kerberos stack in cloud to test a quick change
are close to zero, but I am open to feedback :).

Bug: T268220
Change-Id: If3ab2d71f0a9821d39e21b24e320173d853664e6