cloud: dmz_cidr: detail the list of private production addresses
We don't need to skip the general cloud egress NAT for all 10.x.x.x addresses.
There is a firewall in the core routers that only allows connections to certain
services, the services specified in this change.

This change should not produce any functional behavior change anywhere, but
should help clarify things and be more explicit about the [cloud --> prod priv]
connections.

Change-Id: Ie2f372c780553f55717914670628b7ce8e1927b9
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>