Purge remove apt-listchanges
apt-listchanges was removed in 18ec924831a91ce803abf6004168841b1c4a0178
since it leads to cronspam and doesn't provide much value to us.

It ships a conffile for /etc/apt/apt.conf.d/20listchanges (and those
are retained when removing the package), but then in addition the
/etc/apt/apt.conf.d/ directory is managed via Puppet with recurse/purge=true,
so the 20listchanges is unowned as Puppet sees it and gets removed (but
dpkg still tracks it). To resolve this, we can simply purge the package
which also drops the conffile.

Change-Id: I27edaca4b7dc9f374558a19136ebeab22e6cb069