elasticsearch: Add per-index metrics
To facilitate planning and investigation this patch starts collecting
the per-index statistics elasticsearch provides for a whitelisted set
of indices. Initially this whitelist is a few of our largest indices
which we would like to have more detailed information about.

Somewhat arbitrarily the initial list is the top 10 indices by store
size (primaries + replicas).

Bug: T265931
Change-Id: I94063a6b8493ac6a11126a4abf50f68db5775f29