mariadb-misc: Increase binary log retention
Increase the binary log retention on misc primary dbs (and other
non-multiinstance servers, for which it had already been increased)
with the aim to have a larger buffer for rolling forward a change
after recovering a backup.

This will increase the disk usage, but latest spec dbs should have
more than enough disk space, and the advantages are worth them.
With a 15 day buffer we would have 2 recent weeks of backups to
roll forward.

It requires a master restart or manual application to be effectively
applied.

Change-Id: Ic3436a6f0e498613265f1f0ecc6deeeb605ab3e6