mediawiki: Use python-pil on buster
The Debian pillow packages provided a compatibility layer shipped as
python-imaging with the following description:

| Description-en: Python Imaging Library
| PIL used to provide a PIL.pth file which allowed you to `import Image` but
| this is deprecated.  Use `from PIL import Image` now, or install this
| compatibility package to temporarily get the old behavior back.

This got dropped in Pillow 5 and as such is no longer present in Buster.

I checked the captcha.py script from the ConfirmEdit extension and
it seems to be compatible with the stock PIL, so we can simply use python-pil
on Buster.

Bug: T245757
Change-Id: Ibd5ad47e5e1a7969be491394989143f5f3aa62ed