hive: set session token handling via database for Hadoop test
This setting is needed to be able to then have multiple instances
of the hive metastore. The idea is that hive sessions are not stored
in memory, but on the database, so that multiple daemons can share the
same client statuses.
This turned out to be a problem with Bigtop in the past, but from
quick tests it doesn't seem to be the case anymore.

Bug: T257412
Change-Id: Ia19d77b506d8977bd1f1dc6f3bc0ae28c2f56e7e