codesearch: Adjust write-config job to restart services
As detailed in T268198, the codesearch-write-config systemd timer will now
take over the responsibility of restarting services as it knows when config
has changed, and can reduce the user impact of restarts.

We need to pass --restart to write_config.py and run it as root. Also adjust
the time to 6:00 UTC, which is 7:00 CET and 22:00 PST, so hopefully affects
less users.

This needs Ic6c31e789f911 in labs/codesearch to be useful, but can be merged
ahead of time.

Bug: T268198
Change-Id: Iff1cdf874ce5425db24c3bcbd95a36305ea2112a