k8s::kubelet: Define --logtostderr=true and --v=0
Why:
We decided to have the configuration solely generated by puppet to
remove confusion.

What:
Define --logtostderr=true, --v=0 and --allow-privileged as those are the
settings that kubelet got from the old /etc/kubernetes/config.

Add a parameter $allow-privileged that defaults to true, to allow
changing this in the kubelet

Change-Id: Id915033ab4efaee65f326960482e0bfb04469643