Add Andrew Otto as approval contact for Hadoop and analytics groups
Adding users to a access groups (i.e. groups which grant shell login and/or
sudo rights to a given service) require the signoff of by the group overseeing
the service (typically by an engineering manager). This oversight serves many
purposes; it ensures that the fitting access is applied and that new people
using a service get a proper introduction into using their access.

So far we haven't explicitly written those down, which makes it hard for new SREs
to find the appropriate point of contact (and given that one only does SRE Clinic
Duty 2-3 times per year even for people who are around for a long time.

As such, let's simply track these approval contacts in data.yaml alongside the
access groups.

Add Andrew Otto for Hadoop and related analytics groups, this was
transferred by Nuria when she left.

Change-Id: I953b3bc9dc097d7b48b3a593507c59175caf1cb9