gerrit: move replica IP config from role to host
Gerrit listens on a service IP which is attached to a specific host. We
have the IPs of the primary Gerrit configured in gerrit1001.yaml but the
IPs for the replica were in the role/codfw/gerrit.yaml.  Move those to
the host definition gerrit2001.yaml for consistency.

Hosts: gerrit1001.wikimedia.org, gerrit2001.wikimedia.org
Change-Id: Ieee753889e853c8d2dc77d56eda3e289cb41060e