prometheus::k8s: Support arbitrary clusters
Why:
We want to setup the new codfw staging cluster and the current structure
is showing its limits. Given that with kubeflow powered cluster (lift
wing) coming, it's prudent to create a more extensive and structured
profile class

The idea is to stop relying on the global k8s_infrastructure_users
variable that ties all k8s clusters to the exact same authentication
method and tokens. That will down the line allow each cluster to define
it's own set of tokens if required. Further down the line, if we manage
to move the authentication to a webhook, it is a given that prometheus
will require to somehow have the token, so this change already is a step
towards that direction

What:
* Create a new hiera lookup, called k8s_clusters. That is a hash that
contains the name and various configuration options for the clusters we
want. Repeat the class for all clusters defined in that hash
* The assumption that those clusters are per DC remains and for this
reason, we if guard it in an $enabled member of the hash
* Stop using the $users lookup
* Add a profile::prometheus::kubernetes::cluster_tokens Hash that will
live in the private repo and just contain those secrets
* Stop including ::profile::prometheus::k8s::staging
* Delete ::profile::prometheus::k8s::staging
* Lint the profile class a bit (spaces etc)
* Alert for all the cluster if the prometheus cache isn't updating

Change-Id: If04de47f9f3e42dc08c8abd2f2a11360b1cbc297