delete the drac module
This is very old from pre-2013
as you can see in git history it was 2013
when Matanya converted it to a module.

After this we had different scripts to change
passwords on DRACs (module/mgmt).

And after those we now have a cookbook.
https://wikitech.wikimedia.org/wiki/Management_Interfaces#cookbook

So we can probably delete the module before we worry
about things like I295b5c1d1701b.

Change-Id: Ie4871ff452631520515a0bca7b50f12e4fd6b335