wmcs_wheel_of_misfortune: avoid race condition with proc info
Depending on how fast everything happens, psutil can be unable to get
the name of the process it is killing to send the emails.

Save that aside before running the kills so the emails go out. They do
sometimes, but sometimes it just hits an exception because the info is
cleaned up already by the OS (go figure).

Bug: T266300
Change-Id: Ic8c27ec4a5dec1ddfe4e6ad1a374a9d953bd9552