wikireplicas: extend maintain_dbusers to multiinstance replicas
This is functionally a noop until the debug material is removed and
the new hash is merged with the legacy_hosts. Unfortunately,
something is ending up with a key that is an array, so that needs to be
found first. I hope to see the output of a puppet run so that I can find
what is done wrong here. Naturally, if there's a key that is an array,
the ordered_yaml function blows up (and it's wrong anyway).

I believe the answer will be largely obvious when I see what it
constructs.

Bug: T268312
Bug: T260389
Change-Id: I11d45c8511de4a98013b8c8a8be162c864bed3da