Move s-nail (providing mail(1)) to standard packages
We have plenty of roles which already use mail(1), add it to standard packages,
it uses little disk space and is useful in general to rely on.

For reasons beyond me there are three different implementation of this very
simple tool (which are managed via Debian alternatives). I picked s-nail since
it's the most actively maintained, thus existing cases in Puppet which use
heirloom-mailx are being migrated to it.

Bug: T268725
Change-Id: Ibf320728e607d8317bff220ae36fbd92faab0b11