Move bsd-mailx (providing mail(1)) to standard packages
We have plenty of roles which already use mail(1), add it to standard packages,
it uses little disk space and is useful in general to rely on.

For reasons beyond me there are three different implementation of this very
simple tool, but only bsd-mailx provides the standard mail(1) alias (due to
slight differences in the options), so using that one.

Bug: T268725
Change-Id: Ib968a1dcda93069af9f885f47bf29a4ec48ce62d