wikistats: use script and separate config to dump mysql from timer
The cron job this systemd timer replaced in I44ac35f66ca3ac4
used redirection of stdout and && to compress the result with gzip.

Both things are not supported directly in a systemd commandline.

So we need to use a shell script. But we also don't want to generate
shell scripts from erb templates, so we need a config file for that
as well if we want to keep the parameters flexible.

This means a simple cron turns into:

- foo.timer which runs..
- foo.service which runs ..
- script redirecting output which reads ..
- config file generated by puppet from erb template

:p

Change-Id: Iffd53199d15da9fc50c8ca1e0904eedc909704bf