wdqs: Switch lag metric to be a gauge
TWO REASONS FOR THIS CHANGE
(1) Recent prometheus changes force counters to be suffixed in _total
which is breaking this metric's consumers. Switching to gauge allows us
to avoid needing to do a
s/blazegraph_lastupdated/blazegraph_lastupdated_total
(2) Technically speaking, a timestamp should be a gauge anyway, since
counters represent a relative (as opposed to absolute) value.

Bug: T269204
Change-Id: I7cd2a7eb3e22147fd2a27af49bf4f3e3d62bfed0