k8s_infrastructure_users: Amend to support groups, avoid uid conflicts
Why:
We already found 1 conflict where a uid was reused. We don't really
track or rely those for now, it's pointless to needlessly track them

What:
* Switch from the <%=,<%- ERB syntax to a more Ruby like syntax for
readbility and output consistency reasons
* Allow passing groups as a member in the struct and use that if
specified instead of type. Allow >1 groups by double quoting it as per
docs
* Switch to using the name of the user everywhere. In our case it's
still guaranteed to be unique so we are going to be ok until we migrate
to a better authentication medium
* For the time being, use the numeric id on old clusters and the name
of the user on new ones (so we don't have to test the migration).

Bug: T269461
Change-Id: I00e813d7418ee09a128fa1ebfc1da4dbc81a73ec