alerting: Disable screen/tmux monitoring on orchestrator hosts
There seems to be an ongoing "Long running screen/tmux" alert on
dborch1001. This is not a problem, many times, screens are needed
for one-time data loads or other expensive operations by DBAs and
other data persistence members. Add the newly created role of
orchestrator to the list of hosts to ignore long running screen/
tmux sessions as a sign of something going badly, to increase
signal-to-noise ratio.

This is similar to the hiera key on existing cumin, mariadb, and
db inventory roles.

Bug: T265990
Change-Id: If9a494b2cfbb6b024521247f4b43a0d70fe3a359