categories: fix prom exporter's broken namespace
Rather than only talking to one of the two blazegraph instances directly
(categories => 9990, wdqs => 9999),
the blazegraph exporter now talks to nginx for the sparql endpoint,
while talking to blazegraph directly for other metrics.

The purpose of using nginx for the sparql endpoint is to take advantage
of the aliase.map functionality that exists at the nginx level.

Bug: T269872
Change-Id: I72274ffe845d3e9d9fed27c04f98feb0f665f6b2