presto: reduce the max heap size from 110G to 100G
I just noticed that presto on an-presto1003 was using, between shared
and heap size, all the 128G available and puppet was not able to run
properly. I also noticed alerts related to the OOM killer acting when
some queries were causing the heap size to expand too much on an-presto1001,
so I think that the current set up is too aggressive. Let's reduce
the heap size to 100G to see if things improve.

Change-Id: I9fc1884cae87516b45ffd6f417217dda67b05c34