etcd: make snapshot interval configurable
In Toolforge, we use lower-performance VM systems for etcd. Etcd is
absolutely pathetic when faced with the slightest performance issue.

Since we have persistent iowait, one thing we might be able to do is
trade load on the system for lower disk usage by messing with the interval
for snapshot writes.

Additionally, we are using an old standard. The default is now 100,000
which is not documented everywhere.
(https://github.com/etcd-io/etcd/commit/c5a9d548358f64483b9fc1726f1a64722c4cdf6f)

Bug: T267966
Change-Id: I5293273d968adb593591a4c85540e34d9d638360