httpbb: add tests for parsoid servers
Translating what Subbu provided on T268524#6682537
to httpbb tests.

Getting the "Hospet" site on en.wiki and as suggested
similar for some other languages, de, es and it.

The first URL will always be a 302 and then the second
one will be 200 with actual content.

[deploy1001:~] $ httpbb --hosts parse2001.codfw.wmnet /home/dzahn/test_parse.yaml
Sending to parse2001.codfw.wmnet...
PASS: 8 requests sent to parse2001.codfw.wmnet. All assertions passed.

The part we don't have here is the "compare if the content
is the same on host A vs host B" but we check that "data-parsoid" is actually in the
body.

Bug: T268524
Change-Id: Ie93e85c479b2f5e03aec1b549c059a277e6054e0