profile::swap: allow to override the kerberos credential cache location
This change should allow to set the KRB5CCACHE environment variable for
all the ephemeral systemd unit created by jupyterhub when a user logs in.
The idea is to force all the jvm-based tools, that don't read /etc/krb5.conf,
to check the user kerberos credential cache in a specific location that it is
not the default one (/tmp/krb5_$id-of-the-user).
This should allow to bypass the PrivateTmp restriction that all the jupyterhub
units have, and allow a user to kinit either in the jupyter terminal or within
a ssh/bash session, not in both like it happens now.

Bug: T255262
Change-Id: I1a3c1e82528e3f92eb4e853a8fcf1bc47dc769d3