doc: allow changing WMF_DOC_PATH from Hiera
In order to use profile::doc on WMCS, we need a way to change
WMF_DOC_PATH from hiera.  Doing that via a define which lets us keep the
Apache configuration to be a file rather than a template (useful for
local hacking as well).

Set the published doc path on doc.devtools to point to
integration/docroot development files.

Change-Id: Id52e0b48cf71018e101468116668cbc7607b17cf