calico-typha: Allow 5473 from kubernetes nodes
Why:
calico-typha runs as a hostNetwork pod as it's a very early pod that
runs way before any kind of networking is ready. calico-node pods will
connect to it and for that to happen we need to have a ferm rule

What:
Add a ferm rule for typha. Reuse $bgp_peers for this one as for now all
kubernetes nodes in the same cluster need to be in that construct. Later
on, if and when we move to a more layered BGP peering hierarchy,
revisit. Leave comment for that

Change-Id: Ideade45c0434e0ecd6eb1633366ec6a99ee9be12