toolforge: upgrade sssd patch version
All toolforge hosts that use apt-pinning and are also buster ended up
with broken puppet last week because of an upgrade that apparently
violated the pin. K8s hosts are not affected because they don't use
the pinning module. The upgraded packages cannot be rolled back, so
the only way is forward!

Bug: T270128
Bug: T270129
Change-Id: I9c6417877609b64de483742ba6d0b6bcf60930d2