lldp: add new per interface neighbours fact
Some interface specificity frack see multiple neighbours on there
interface.  this is uncommon but is supported, as such add an
additional per interface `neighbours` fact.

We keep the old per interface neighbour fact for backwards
compatibility.  It is set to the last interface listed in
the output of `lldp -f xml`

New output looks like the following

  $ sudo facter -p lldp
  {
    eno1 => {
      neighbors => [
        "asw2-d-eqiad"
      ],
      neighbor => "asw2-d-eqiad",
      port => "ge-6/0/5",
      descr => "sretest1002:eno1 {#}",
      mtu => 9192,
      vlans => {
        untagged_vlan => 1020,
        mode => "access"
      }
    }
  }

Bug: T268802
Change-Id: I11d4f383112ebe1341113f11cba64fc9888bdd08