icinga/prometheus: add notes_link to puppetboard to puppet failure alerts
Currently when this alert triggers we have output like this on IRC:

<snip>
<+icinga-wm> PROBLEM - Widespread puppet agent failures on alert1001 is
CRITICAL: 0.01183 ge 0.01
https://wikitech.wikimedia.org/wiki/Monitoring/Missing_notes_link
https://grafana.wikimedia.org/d/yOxVDGvWk/puppet
</snip>

It displays the "Missing_notes_link" because it's.. missing a notes link.

It does show a Grafana link in addition.

When you follow that from there you also get this puppetboard link but
it is trying to embed it and that doesn't work for
me in Firefox.

So I have to click again to breakout to a new tab for that.

Instead we should fix the error in IRC output and link to the puppetboard
URL right away..because that's where people see the interesting
information; the actual names of hosts with broken puppet.

Bug: T270354
Change-Id: I5631412bf85276c52a79cde4a95a57b16734aff0