uwsgi: remove postrotate step to avoid reloads after logrotation
It happens daily that when logrotate runs on some ores nodes, the
reload issued (that eventually ends up in a SIGHUP) causes a shutdown
of all uwsgi workers (expected) but then they are not brought up again
(the master process probably dies as well). Then ores-uwsgi stays down
until the next puppet run.
I am wondering if a reload is needed if we use copytruncate, so I opened
this change to start the discussion :)

Change-Id: If11201a28acc7350c5597259d5890fb79311a2da