puppetmaster: remove default apache ports from puppetmaster
We don't want apache to listen on the default ports for puppetmaster.
The puppetmaster process listens on 8140 and 8141.  And the plan would
be to perform TLS termination via envoy so we don't want apache
listening on 443

config-master runs on port 80 so an explicit configuration in that
profile

Bug: T263831
Hosts: re:puppetmaster.*wmnet$
Change-Id: I7ff95933fbc5ebf37dea31026f9a2517dd910438