cumin: Don't complain if A:spare matches no hosts.
check_cumin_aliases fires an alert if any alias matches zero hosts,
since that's normally a configuration error. But A:spare (which is just
O:spare::system) will sometimes be empty if there just don't happen to
be any spare hosts at any given moment. That's fine, and the alias is
still useful; nothing's wrong, so we shouldn't alert.

Change-Id: I2894a0de95fc606521b2d63eb34570ec515e077d