profile: add priority to logstash filter filenames
logstash-filter-verifier cannot prove the effects of logstash configuration
on events when the filters are applied out of order.
rename the filters and amend the Makefile to take advantage of the
priority ordering.

amend broken tests.
update owner/group to logstash for consistency

should be largely a noop

Hosts: O:logstash,O:logstash7
Bug: T254533
Change-Id: Id1a7d2bd59f6ea507aea46ba74e8aebd37542eb8