cloud haproxy: refactor the various haproxy setups
This also abstracts the base haproxy install into a little corner of the haproxy
top level module in a way that should be flexible enough for all VM-based
instances without worrying about clashes.

Since puppet only needs one of notify or subscribe, subconfigs are easy
to place in the profiles.

Change-Id: I2c8c6a566abcc188eae6d25832071d6bce533933