wmcs.backup: Remove all dangling snapshots
This remove any dangling snapshot for the vms already being backed up
in this host.

As we only need one snapshot per VM/image, we can remove any other
snapshots.
This snapshot must be matched with a local backup, otherwise it's not
useful and we can remove it too to force the creation of a new one
next time.

Bug: T267195
Change-Id: Id0c7c24b32d5ff3ed9fb1eea931fbf013d8a6e35
Signed-off-by: David Caro <dcaro@wikimedia.org>