install_server: add a "rescue" label
This change should add an option to PXE boot in Debian Rescue mode,
leaving the default to "server" (and hence install) as it is now.
I didn't find a way to force a specific node to use the new label
in the dhcp config, so I guess this happens manually when PXE starts
(but I am not completely sure about it, please correct me if I am wrong).

The use case is to execute grub-install on /dev/sdb on an-coord1002,
since a disk died (/dev/sda) and the mbr is not present on the sdb disk
(making the host unbootable even if one disk of the raid 1 array works).
I tried to use the rescue shell in regular d-i, but I don't see any
"real" partition to chroot to under /target (nor anything useful to mount
under /dev).

Bug: T191018
Change-Id: I9d05be9b473b6e3260da238cb9539fcc26f5eace