mw maintenance: Install php-readline from component/php72
This was already an issue with Stretch, but the problem didn't show up
since PHP 7.2 has a higher version than what's in default Stretch (7.0).
But with Buster we need to explicitly select the 7.2 version from the
component, otherwise we end up with the 7.3 version instead of 7.2

Also move away from require_package in this class.

Bug: T245757
Bug: T266479
Change-Id: Icc5871ee6cfd80a01499fd9f38f9ad075b93ec22