wmcs.backup: Add backup_image command
This allows backing up an image, cleaning up after if necessary.

== Test plan

Ran it twice on cloudcontrol1003, and it did both backups removing the
old snapshot after and keeping only the latest one:
https://phabricator.wikimedia.org/P13650

Then ran it twice also on cloudcontrol1005 (that also backs up images)
and saw that it generated the new snapshots too (based on the old
ones) and kept only the last one:
https://phabricator.wikimedia.org/P13651

Interesting enough it was way faster on cloudcontrol1005 (as in 10min
vs 10sec), not sure why though.

Bug: T270478
Change-Id: I6aa6bd36015dffad017eecaafc6f50c77015c513
Signed-off-by: David Caro <dcaro@wikimedia.org>