admin: remove members of 'reseachers' already in other posix groups
The users removed from the 'researchers' group (to be deprecated)
didn't answer to several pings (via email and phabricator) and are
already in other posix groups (so we don't need to remove ssh access).

This is a change to reduce the number of users in 'researchers' while
working on the other accounts.

Bug: T268801
Change-Id: Icdfcf241e54d5a8d1f05b0d0d0f2cd487572aaab