parsoid::testing: fix duplicate declaration of mariadb-client for buster
In the role that uses this profile on buster we are now including a
generic mariadb profile which already pulls this package.

Having it here as well causes a duplicate declaration. scandium
on stretch should not be affected though.

The client config also needs to be moved so it only applies
to stretch hosts or it will also conflict with the default config
from the generic profile.

Bug: T266509
Change-Id: I0176f595892775646edcdcffe350b8c6fe354c15