Use $region for default mcrouter routes
This is more intuitive and makes it easier to send test commands to DCs.

Note that MediaWiki is not configured to specify the regions in keys; the
route is ommitted on DC-local reads/writes and a wildcard region is used on
cross-DC writes. If, in the future, explicit routes are ever needed on reads,
then it will use the local DC route anyway.

The default region is $::site, so this does not effect MediaWiki traffic.

Change-Id: Ie3796b01e05d1c4e860f37c3e3186c01d84a14ed