cloud.encapi: enable ssl nginx vhost
This makes nginx listen on a new port (ip4 and 6) using https protocol
to serve the same content as it servers on the default http port.
This allows to have both to start migrating hosts over to https.

Note: This also moves the labs_networks variable as a parameter to be
      able to pass it, as it has a class internal variables as default
      it was hard to mock/customize for the tests.

== Test plan
Created manually the config (a new one) on root@cloud-puppetmaster-03
with the same configuration and reloaded nginx, then tried executing
a request to that new port seeing that it worked well using http:

```
root@cloud-puppetmaster-03# curl \
    https://cloud-puppetmaster-03.cloudinfra.eqiad.wmflabs:8143\
        /v1/maps/node/maps-puppetmaster.maps.eqiad.wmflabs

hiera: {
    mount_nfs: true,
    'role::puppetmaster::standalone::autosign': 'true'
}
roles: ['role::puppetmaster::standalone']
```

Bug: T268877
Bug: T266479
Change-Id: I47b2e2c78969c175840e89830ba4a5667de0791e
Signed-off-by: David Caro <dcaro@wikimedia.org>