wikireplicas: add a multiinstance role for the dedicated analytics host
In order to clear the way for rebuilding labsdb1012 as a multiinstance
host it will need a new role so that it includes the profile that adds
the firewall rule needed by analytics team as well as keeping it separate
from the cloud proxies. Please note that puppet *will* register a failure
because of T260511. Puppet will still finish a run, but you'll see red
text.

Bug: T269211
Change-Id: I12c04a33dc02b9876ed0a3d0b629e970dbbc80ee