cloudweb mcrouter config: use AllFastestRoute for write actions
With AllSyncRoute then Horizon errors out anytime it's unable
to write to either host. That ensures data integrity but also
means that we get downtime if either labweb is down.

With AllFastestRoute there's a small chance of data inconsistency
(if one host goes down, then comes back, then the other goes down)
but I can live with that.

Bug: T271349
Change-Id: I90472cee02b22cf49fbdf7dd5fced02bc144ddde