maintain-meta_p: simplify getting a str from requests module
With requests of any recent vintage, we can sidestep the issue of the
content attribute returning something other than unicode back in Debian
Stretch days by using the json() convenience function and drop one more
import. Tested on Stretch, and it returns a json full of str (which
we guard against even if a field or two end up as bytes somehow).

This should fix the request block failing for new wikis on legacy servers.

Bug: T270839
Change-Id: Ic906ac5c13729f139a98df5d47d918df9f465479