docker_registry_ha: Add a script to generate a static HTML homepage
This is a rewrite of <https://github.com/thcipriani/dockerregistry> to
fully be in Python and avoid bugs from edge cases (e.g. removed images
or
images with only one tag).

A systemd timer will run the script hourly, which generates a tree of
HTML
files for nginx to serve.

Bug: T179696
Change-Id: Idd878538973db8efeae8e0ad9d2eb11b55ef6780