Horizon: Disable offline compression in Train
I have a strange issue that appears on a production web
server (but /not/ on the test server.) When deploying, the
final compression stage fails like this:

  CommandError: An error occurred during rendering \
  /site-packages/openstack_dashboard/templates/horizon/_scripts.html: \
  Couldn't find any precompiler in COMPRESS_PRECOMPILERS setting for \
  mimetype '\'text/javascript\''.

That's the result of some strange double-quoting issue. It does not seem to affect the actual
running service. Rather than spend another day chasing this bug I'm going to disable offline
compression in Train and hope that we can switch it back on in a future version.

Bug: T269004
Change-Id: I6a4d0fc8f7c76f272a758a813f7391a2c4d15187