Add a wrapper to optionally mail the output of a systemd timer
For some timers we need the equivalent of the old cron behaviour to send
an email if an executed command produces output (e.g. the check of
Cumin aliases).

I tried to experiment with a few systemd native mechanisms (such as a
multi instance systemd-mail@.service which sends a mail via mailx
which gets spawned by an OnFailure directive in the unit), but it's
rather intricate and a wrapper turned out to be simpler to integrate
in the end.

A followup patch will amend systemd::timer::job with an option to
invoke the wrapper where it's needed.

Change-Id: I753aac5742e0010ed47a36a67dd54b1a3e5dffb0